refactor(dashboard): Co-locate DashboardPlugin in dashboard package
