--- packages/dashboard-plugin/package.json
@@ -1,40 +1,0 @@
-{
-    "name": "@vendure/dashboard-plugin",
-    "version": "3.3.7",
-    "main": "dist/index.js",
-    "types": "dist/index.d.ts",
-    "files": [
-        "dist/**/*"
-    ],
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/vendure-ecommerce/vendure"
-    },
-    "license": "GPL-3.0-or-later",
-    "scripts": {
-        "build": "rimraf dist && tsc -p ./tsconfig.build.json",
-        "watch": "tsc -p ./tsconfig.build.json --watch",
-        "lint": "eslint --fix .",
-        "compile": "tsc -p ./tsconfig.build.json"
-    },
-    "homepage": "https://www.vendure.io",
-    "funding": "https://github.com/sponsors/michaelbromley",
-    "publishConfig": {
-        "access": "public"
-    },
-    "devDependencies": {
-        "@types/express": "^5.0.1",
-        "@types/fs-extra": "^11.0.4",
-        "@vendure/common": "3.3.7",
-        "@vendure/core": "3.3.7",
-        "@vendure/dashboard": "3.3.7",
-        "express": "^5.1.0",
-        "rimraf": "^5.0.5",
-        "typescript": "5.8.2"
-    },
-    "dependencies": {
-        "date-fns": "^4.0.0",
-        "express-rate-limit": "^7.5.0",
-        "fs-extra": "^11.2.0"
-    }
-}
