--- packages/dashboard-plugin/src/index.ts
@@ -1,2 +1,0 @@
-export * from './plugin';
-export * from './types';
