--- packages/dashboard-plugin/tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "composite": false,
-        "incremental": false,
-        "paths": {}
-    },
-    "include": ["src/**/*"],
-    "exclude": ["**/*.spec.ts", "e2e/**"]
-}
