--- packages/dashboard-plugin/tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "baseUrl": ".",
-        "rootDir": "./src",
-        "outDir": "./dist",
-        "declaration": true,
-        "removeComments": true,
-        "strictPropertyInitialization": false,
-        "sourceMap": true
-    }
-}
