+++ packages/dashboard/plugin/api/metrics.resolver.ts
@@ -1,8 +1,8 @@
 import { Args, Query, Resolver } from '@nestjs/graphql';
 import { Allow, Ctx, Permission, RequestContext } from '@vendure/core';
 
-import { MetricsService } from '../service/metrics.service';
-import { MetricSummary, MetricSummaryInput } from '../types';
+import { MetricsService } from '../service/metrics.service.js';
+import { MetricSummary, MetricSummaryInput } from '../types.js';
 
 @Resolver()
 export class MetricsResolver {
