+++ packages/dashboard/plugin/index.ts
@@ -1,0 +1,2 @@
+export * from './dashboard.plugin.js';
+export * from './types.js';
